fix result message
Previously, the operation counts in the experiment result message were
inconsistent. This is because we maintain two separate operation
counts---first in a measurements object, and second in the benchmark
`OpsCounts`.

This is confusing, and we should eventually centralize the counts, but
one immediate problem is that the post-warmup operation count reset does
not work as intended. We only reset the measurements object, but not the
`OpsCounts`. This causes a report mismatch in the experiment result
message.

This commit fixes this by using the measurements object count for the
total operation counts in the result message. Followup work is needed to
correctly reset operation counts when the warmup phase ends.